fixed typing issue

+++ gps_demo/gpsSensor.h
@@ -1,6 +1,7 @@
 #ifndef GPS_SENSOR_H
 #define GPS_SENSOR_H
 
+#include <Arduino.h>
 #include <Wire.h> //Needed for I2C to GPS
 #include "SparkFun_Ublox_Arduino_Library.h" //http://librarymanager/All#SparkFun_Ublox_GPS
 #include <MicroNMEA.h> //http://librarymanager/All#MicroNMEA
@@ -19,9 +20,9 @@ bool processIncomingData();
 bool gpsDataValid();
 
 // Returns latitude
-long getLatitude();
+double getLatitude();
 
 // Returns longitude
-long getLongitude();
+double getLongitude();
 
 #endif
